+++ alacritty/src/window.rs
@@ -23,7 +23,10 @@ use glutin::event_loop::EventLoop;
 #[cfg(target_os = "macos")]
 use glutin::platform::macos::{RequestUserAttentionType, WindowBuilderExtMacOS, WindowExtMacOS};
 #[cfg(not(any(target_os = "macos", windows)))]
-use glutin::platform::unix::{EventLoopWindowTargetExtUnix, WindowBuilderExtUnix, WindowExtUnix};
+use glutin::platform::unix::{
+    ButtonState as WaylandThemeButtonState, EventLoopWindowTargetExtUnix, Theme as WaylandTheme,
+    WindowBuilderExtUnix, WindowExtUnix,
+};
 #[cfg(not(target_os = "macos"))]
 use glutin::window::Icon;
 use glutin::window::{CursorIcon, Fullscreen, Window as GlutinWindow, WindowBuilder, WindowId};
@@ -35,6 +38,8 @@ use x11_dl::xlib::{Display as XDisplay, PropModeReplace, XErrorEvent, Xlib};
 
 use alacritty_terminal::config::{Decorations, StartupMode, WindowConfig};
 use alacritty_terminal::event::Event;
+#[cfg(not(any(target_os = "macos", windows)))]
+use alacritty_terminal::term::color::Rgb;
 #[cfg(not(windows))]
 use alacritty_terminal::term::{SizeInfo, Term};
 
@@ -361,6 +366,27 @@ impl Window {
         self.window().wayland_display()
     }
 
+    #[cfg(not(any(target_os = "macos", target_os = "windows")))]
+    pub fn set_wayland_theme(&mut self, config: &Config) {
+        let hovered_close_icon = config.colors.normal().red;
+        let hovered_maximize_icon = config.colors.normal().green;
+        let hovered_minimize_icon = config.colors.normal().yellow;
+        let foreground = config.colors.primary.foreground;
+        let background = config.colors.primary.background;
+        let dim_foreground =
+            config.colors.primary.dim_foreground.unwrap_or_else(|| foreground * 0.66);
+
+        let theme = AlacrittyWaylandTheme {
+            foreground,
+            background,
+            dim_foreground,
+            hovered_close_icon,
+            hovered_minimize_icon,
+            hovered_maximize_icon,
+        };
+        self.window().set_wayland_theme(theme);
+    }
+
     /// Adjust the IME editor position according to the new location of the cursor
     #[cfg(not(windows))]
     pub fn update_ime_position<T>(&mut self, terminal: &Term<T>, size_info: &SizeInfo) {
@@ -383,6 +409,93 @@ impl Window {
 
     fn window(&self) -> &GlutinWindow {
         self.windowed_context.window()
+    }
+}
+
+#[derive(Debug, Clone)]
+struct AlacrittyWaylandTheme {
+    background: Rgb,
+    foreground: Rgb,
+    dim_foreground: Rgb,
+    hovered_close_icon: Rgb,
+    hovered_maximize_icon: Rgb,
+    hovered_minimize_icon: Rgb,
+}
+
+impl WaylandTheme for AlacrittyWaylandTheme {
+    fn primary_color(&self, _window_active: bool) -> [u8; 4] {
+        [0xff, self.background.r, self.background.g, self.background.b]
+    }
+
+    fn secondary_color(&self, window_active: bool) -> [u8; 4] {
+        if window_active {
+            [0xff, self.foreground.r, self.foreground.g, self.foreground.b]
+        } else {
+            [0xff, self.dim_foreground.r, self.dim_foreground.g, self.dim_foreground.b]
+        }
+    }
+
+    fn close_button_color(&self, _status: WaylandThemeButtonState) -> [u8; 4] {
+        [0x00, self.background.r, self.background.g, self.background.b]
+    }
+
+    fn close_button_icon_color(&self, status: WaylandThemeButtonState) -> [u8; 4] {
+        match status {
+            WaylandThemeButtonState::Hovered => [
+                0xff,
+                self.hovered_close_icon.r,
+                self.hovered_close_icon.g,
+                self.hovered_close_icon.b,
+            ],
+            WaylandThemeButtonState::Idle => {
+                [0xff, self.foreground.r, self.foreground.g, self.foreground.b]
+            },
+            WaylandThemeButtonState::Disabled => {
+                [0xff, self.dim_foreground.r, self.dim_foreground.g, self.dim_foreground.b]
+            },
+        }
+    }
+
+    fn maximize_button_color(&self, _status: WaylandThemeButtonState) -> [u8; 4] {
+        [0x00, self.background.r, self.background.g, self.background.b]
+    }
+
+    fn maximize_button_icon_color(&self, status: WaylandThemeButtonState) -> [u8; 4] {
+        match status {
+            WaylandThemeButtonState::Hovered => [
+                0xff,
+                self.hovered_maximize_icon.r,
+                self.hovered_maximize_icon.g,
+                self.hovered_maximize_icon.b,
+            ],
+            WaylandThemeButtonState::Idle => {
+                [0xff, self.foreground.r, self.foreground.g, self.foreground.b]
+            },
+            WaylandThemeButtonState::Disabled => {
+                [0xff, self.dim_foreground.r, self.dim_foreground.g, self.dim_foreground.b]
+            },
+        }
+    }
+
+    fn minimize_button_color(&self, _status: WaylandThemeButtonState) -> [u8; 4] {
+        [0x00, self.background.r, self.background.g, self.background.b]
+    }
+
+    fn minimize_button_icon_color(&self, status: WaylandThemeButtonState) -> [u8; 4] {
+        match status {
+            WaylandThemeButtonState::Hovered => [
+                0xff,
+                self.hovered_minimize_icon.r,
+                self.hovered_minimize_icon.g,
+                self.hovered_minimize_icon.b,
+            ],
+            WaylandThemeButtonState::Idle => {
+                [0xff, self.foreground.r, self.foreground.g, self.foreground.b]
+            },
+            WaylandThemeButtonState::Disabled => {
+                [0xff, self.dim_foreground.r, self.dim_foreground.g, self.dim_foreground.b]
+            },
+        }
     }
 }